+++ changelog.md
@@ -25,6 +25,7 @@ Codion Change Log
 - Ancestor added.
 - Utilities.systemLookAndFeelClassName() moved to LookAndFeelEnabler, systemOrCrossPlatformLookAndFeelEnabled() removed.
 - FilterTableCellRenderer.Builder.toolTipData() replaced with toolTip() function.
+- FilterTableCellRenderer.Configurer added along with Builder.configurer(), simplified and refactored.
 ### is.codion.framework.domain
 - EntityType.define() and AttributeDefiner.define() renamed as().
 - DerivedAttributeDefinition.sources() renamed attributes().